backend/requirements.txt
Django==5.0.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
paystack==1.5.0
Pillow==10.2.0
django-filter==24.1
djangorestframework-simplejwt==5.3.1
python-magic==0.4.27